[Fix] AgendaMockData 병합 및 빌더패턴 추가
